Initial commit with code, readme and license
Initial commit with code, readme and an added license as specified on the source website: https://figshare.com/s/254feb5ce70a340b0d57?file=14702084 on 3/9/21; which is the supplementary material of the paper: 2019 ``The influence of invariant solutions in the transient behaviour of an air bubble in a Hele-Shaw channel", Keeler, J. S., Thompson, A. B. & Lemoult, G. & Juel, A. & Hazel, A. H., Proc. R. Soc. A, 20190434.